pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blux-quantum"
version = "0.1.0"
description = "BLUX Quantum — Unified Operator CLI (bluxq) for the BLUX ecosystem"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Apache-2.0" }
authors = [{ name = "BLUX", email = "ops@blux.systems" }]
keywords = ["cli", "security", "observability", "tui", "blux"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Topic :: Security",
]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.0",
  "textual>=0.62.0",
  "psutil>=5.9.8",
  "pyyaml>=6.0.1",
  "sqlite-utils>=3.36",
  "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
dev = ["pytest", "mypy", "ruff", "pre-commit", "mkdocs-material"]

[project.scripts]
bluxq = "blux_quantum.cli:app"

[project.entry-points."blux.plugins"]
guard = "blux_quantum.plugins.examples.guard_plugin:get_app"
lite = "blux_quantum.plugins.examples.lite_plugin:get_app"
doctrine = "blux_quantum.plugins.examples.doctrine_plugin:get_app"

[tool.ruff]
target-version = "py39"
line-length = 100
select = ["E", "F", "I", "UP", "B", "A"]
ignore = ["B905"]
